requirements.txt
django
djangorestframework
psycopg2-binary
Pillow
python-decouple
djangorestframework-simplejwt
drf-yasg

